Added city visuals and reorganized code
